Separa pid_provider.controller em dois base_pid_provider e controller para manter compatíveis core e upload
